- refactor : enum 패키지 이동 및 컨벤션 수정
