+++ src/main/java/racingcar/Application.java
@@ -9,8 +9,10 @@ import racingcar.view.GameView;
 import java.util.List;
 
 public class Application {
+
     public static void main(String[] args) {
         RacingGame racingGame = getRacingGame();
+        GameView.printStartMessage();
         racingGame.start();
     }
